refactor(storage): organize durable engine modules
