+++ crates/astrid-storage-engine/src/durable/faults.rs
