+++ crates/astrid-storage-engine/src/durable/validation.rs
